+++ packages/backend/tests/convexTest/reminders.test.ts
@@ -1,0 +1,166 @@
+/**
+ * Reminder Flow Tests
+ */
+
+import { convexTest } from "convex-test";
+import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
+import { internal } from "../../convex/_generated/api";
+import schema from "../../convex/schema";
+import { modules } from "../test.setup";
+import { mockTelegramResponse } from "./fixtures/testHelpers";
+
+let sentMessages: { chatId: string; text?: string }[] = [];
+
+function setupFetchMock() {
+	sentMessages = [];
+
+	vi.stubGlobal(
+		"fetch",
+		vi.fn(async (url: string, options?: RequestInit) => {
+			// Mock Telegram sendMessage
+			if (url.includes("api.telegram.org") && url.includes("/sendMessage")) {
+				let body: any = {};
+				if (options?.body instanceof FormData) {
+					body = Object.fromEntries(options.body as any);
+				} else if (typeof options?.body === "string") {
+					body = JSON.parse(options.body);
+				}
+				sentMessages.push({ chatId: body.chat_id, text: body.text });
+				return {
+					ok: true,
+					json: async () => mockTelegramResponse(),
+				} as Response;
+			}
+
+			// Mock Convex storage
+			if (url.includes("convex.cloud") || url.includes("convex.site")) {
+				return {
+					ok: true,
+					arrayBuffer: async () => new ArrayBuffer(100),
+					blob: async () => new Blob(["voucher-image"], { type: "image/jpeg" }),
+				} as Response;
+			}
+
+			console.warn(`Unmocked fetch: ${url}`);
+			return { ok: false, status: 404 } as Response;
+		}),
+	);
+}
+
+describe("Reminder Flow", () => {
+	beforeEach(() => {
+		setupFetchMock();
+		vi.stubEnv("TELEGRAM_BOT_TOKEN", "test-bot-token");
+	});
+
+	afterEach(() => {
+		vi.unstubAllGlobals();
+		vi.unstubAllEnvs();
+	});
+
+	test("sends reminders to users who claimed vouchers yesterday", async () => {
+		vi.useFakeTimers();
+
+		const now = Date.now();
+		const oneDayMs = 24 * 60 * 60 * 1000;
+		const sevenDaysMs = 7 * 24 * 60 * 60 * 1000;
+		const yesterday = now - oneDayMs;
+		const twoDaysAgo = now - 2 * oneDayMs;
+		const futureExpiry = now + sevenDaysMs;
+
+		const t = convexTest(schema, modules);
+
+		const claimerId = await t.run(async (ctx) => {
+			return await ctx.db.insert("users", {
+				telegramChatId: "claimer123",
+				coins: 100,
+				isBanned: false,
+				createdAt: now,
+				lastActiveAt: now,
+			});
+		});
+
+		const uploaderId = await t.run(async (ctx) => {
+			return await ctx.db.insert("users", {
+				telegramChatId: "uploader456",
+				coins: 50,
+				isBanned: false,
+				createdAt: now,
+				lastActiveAt: now,
+			});
+		});
+
+		// Create voucher claimed yesterday (should trigger reminder)
+		const yesterdayVoucherId = await t.run(async (ctx) => {
+			const imageStorageId = await ctx.storage.store(new Blob(["test"]));
+			return await ctx.db.insert("vouchers", {
+				type: "10",
+				status: "claimed",
+				imageStorageId,
+				uploaderId,
+				claimerId,
+				claimedAt: yesterday,
+				expiryDate: futureExpiry,
+				createdAt: yesterday,
+			});
+		});
+
+		// Create voucher claimed today (should NOT trigger)
+		const todayVoucherId = await t.run(async (ctx) => {
+			const imageStorageId = await ctx.storage.store(new Blob(["test"]));
+			return await ctx.db.insert("vouchers", {
+				type: "5",
+				status: "claimed",
+				imageStorageId,
+				uploaderId,
+				claimerId,
+				claimedAt: now,
+				expiryDate: futureExpiry,
+				createdAt: now,
+			});
+		});
+
+		// Create voucher claimed 2 days ago (should NOT trigger)
+		const oldVoucherId = await t.run(async (ctx) => {
+			const imageStorageId = await ctx.storage.store(new Blob(["test"]));
+			return await ctx.db.insert("vouchers", {
+				type: "20",
+				status: "claimed",
+				imageStorageId,
+				uploaderId,
+				claimerId,
+				claimedAt: twoDaysAgo,
+				expiryDate: futureExpiry,
+				createdAt: twoDaysAgo,
+			});
+		});
+
+		const chatIds = await t.query(
+			internal.reminders.getUsersWhoClaimedYesterday,
+			{},
+		);
+
+		expect(chatIds).toHaveLength(1);
+		expect(chatIds[0]).toBe("claimer123");
+
+		sentMessages.length = 0;
+		await t.action(internal.reminders.sendDailyUploadReminders, {});
+
+		vi.runAllTimers();
+		await t.finishInProgressScheduledFunctions();
+
+		const reminderMessage = sentMessages.find(
+			(m) =>
+				m.chatId === "claimer123" &&
+				m.text?.includes("Upload your new vouchers"),
+		);
+		expect(reminderMessage).toBeDefined();
+
+		const uploaderMessage = sentMessages.find(
+			(m) => m.chatId === "uploader456",
+		);
+		expect(uploaderMessage).toBeUndefined();
+
+		vi.useRealTimers();
+	});
+});
